Split stack-event-log component
