+++ app/assets/javascripts/infrastructures.js
@@ -55,26 +55,7 @@
   var alert_and_show_infra = helpers.alert_and_show_infra;
 
 
-  Vue.component("stack-events-table", {
-    props: { events: Array, },
-    template: '#stack-events-table-template',
-    methods: {
-      event_tr_class: function (status) {
-        if      (status === "CREATE_COMPLETE")    { return "success"; }
-        else if (status.indexOf("FAILED") !== -1) { return "danger"; }
-        else if (status.indexOf("DELETE") !== -1) { return "warning"; }
-        return '';
-      },
-      toLocaleString: toLocaleString,
-    },
-    created: function () {
-      var self = this;
-      console.log(self);
-      this.$watch('events', function () {
-        $(self.$el).hide().fadeIn(800);
-      });
-    },
-  });
+  Vue.component('stack-events-table', require('infrastructures/stack-events-table.js'));
 
   Vue.component("add-modify-tabpane", {
     props: {
